requirements.txt
# Core dependencies
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
pydantic==2.5.2
boto3==1.34.7
openai==1.84.0

# Database
psycopg2-binary==2.9.9

# Document processing
PyMuPDF==1.23.8
PyPDF2==3.0.1
python-dateutil==2.8.2
dateparser==1.2.0

# Utilities
numpy==1.26.2
python-dotenv==1.0.0

# AWS services
botocore==1.34.7

# JSON handling
simplejson==3.19.2

# Text processing
nltk==3.8.1
tiktoken==0.5.2

# CLI and monitoring
rich==13.7.0
click==8.1.7

# Monitoring
# cloudwatch==1.1.2  # Not needed, using boto3 for CloudWatch

# Document processing - Textract
amazon-textract-textractor==1.5.0
pytesseract==0.3.10
pdf2image==1.16.3
